+++ compare_hashes/submission.cpp
@@ -1,7 +1,4 @@
-#include <string>
-#include <unordered_map>
 #include <map>
-#include <unordered_set>
 #include <set>
 #include <vector>
